+++ sync-for-payables/CodatSyncPayables/CodatSyncPayables.cs
@@ -17,6 +17,8 @@ namespace Codat.Sync.Payables
     using System.Threading.Tasks;
     using System;
 
+
+
     /// <summary>
     /// Sync for Payables: The API for Sync for Payables. <br/>
     /// 
@@ -130,6 +132,22 @@ namespace Codat.Sync.Payables
     
     public class SDKConfig
     {
+        public static string[] ServerList = new string[]
+        {
+            "https://api.codat.io",
+        };
+        /// Contains the list of servers available to the SDK
+        public string serverUrl = "";
+        public int serverIndex = 0;
+
+        public string GetTemplatedServerDetails()
+        {
+            if (!String.IsNullOrEmpty(this.serverUrl))
+            {
+                return Utilities.TemplateUrl(Utilities.RemoveSuffix(this.serverUrl, "/"), new Dictionary<string, string>());
+            }
+            return Utilities.TemplateUrl(SDKConfig.ServerList[this.serverIndex], new Dictionary<string, string>());
+        }
     }
 
     /// <summary>
@@ -167,16 +185,12 @@ namespace Codat.Sync.Payables
     public class CodatSyncPayables: ICodatSyncPayables
     {
         public SDKConfig Config { get; private set; }
-        public static List<string> ServerList = new List<string>()
-        {
-            "https://api.codat.io",
-        };
 
         private const string _language = "csharp";
-        private const string _sdkVersion = "3.0.0";
-        private const string _sdkGenVersion = "2.188.1";
+        private const string _sdkVersion = "3.0.1";
+        private const string _sdkGenVersion = "2.194.1";
         private const string _openapiDocVersion = "3.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 3.0.0 2.188.1 3.0.0 Codat.Sync.Payables";
+        private const string _userAgent = "speakeasy-sdk/csharp 3.0.1 2.194.1 3.0.0 Codat.Sync.Payables";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;
@@ -196,9 +210,14 @@ namespace Codat.Sync.Payables
         public ITrackingCategories TrackingCategories { get; private set; }
         public IPushOperations PushOperations { get; private set; }
 
-        public CodatSyncPayables(Security? security = null, string? serverUrl = null, ISpeakeasyHttpClient? client = null)
+        public CodatSyncPayables(Security? security = null, int? serverIndex = null, string? serverUrl = null, Dictionary<string, string>? urlParams = null, ISpeakeasyHttpClient? client = null)
         {
-            _serverUrl = serverUrl ?? CodatSyncPayables.ServerList[0];
+            if (serverUrl != null) {
+                if (urlParams != null) {
+                    serverUrl = Utilities.TemplateUrl(serverUrl, urlParams);
+                }
+                _serverUrl = serverUrl;
+            }
 
             _defaultClient = new SpeakeasyHttpClient(client);
             _securityClient = _defaultClient;
@@ -210,6 +229,7 @@ namespace Codat.Sync.Payables
             
             Config = new SDKConfig()
             {
+                serverUrl = _serverUrl
             };
 
             Companies = new Companies(_defaultClient, _securityClient, _serverUrl, Config);