+++ sync-for-payables/CodatSyncPayables/Models/Operations/CreateAccountResponse.cs
@@ -40,6 +40,6 @@ namespace Codat.Sync.Payables.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
     }
 }