+++ sync-for-payables/CodatSyncPayables/Models/Operations/GetTrackingCategoryResponse.cs
@@ -35,7 +35,7 @@ namespace Codat.Sync.Payables.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
 
         /// <summary>
         /// Success