+++ sync-for-payables/CodatSyncPayables/Models/Webhooks/ClientRateLimitResetResponse.cs
@@ -29,6 +29,6 @@ namespace Codat.Sync.Payables.Models.Webhooks
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
     }
 }